third_party/blink/web_tests/platform/generic/external/wpt/css/mediaqueries/media-query-matches-in-iframe-expected.txt

This is a testharness.js-based test.
PASS matchMedia('(max-width: 150px)').matches should update immediately
PASS matchMedia('(width: 100px)').matches should update immediately
PASS matchMedia('(orientation: portrait)').matches should update immediately
PASS matchMedia('(aspect-ratio: 1/1)').matches should update immediately
PASS matchMedia('(max-aspect-ratio: 4/3)').matches should update immediately
PASS matchMedia('(height: 100px)').matches should update immediately
PASS matchMedia('(max-height: 150px)').matches should update immediately
PASS matchMedia('(min-aspect-ratio: 3/4)').matches should update immediately
PASS matchMedia('(min-height: 150px)').matches should update immediately
PASS matchMedia('(aspect-ratio: 1/2)').matches should update immediately
PASS matchMedia('(min-width: 150px)').matches should update immediately
PASS matchMedia('(min-aspect-ratio: 4/3)').matches should update immediately
FAIL matchMedia('(max-width: 150px)') should not receive a change event until update the rendering step of HTML5 event loop assert_equals: expected 1 but got 0
FAIL matchMedia('(width: 100px)') should not receive a change event until update the rendering step of HTML5 event loop assert_equals: expected 1 but got 0
FAIL matchMedia('(orientation: portrait)') should not receive a change event until update the rendering step of HTML5 event loop assert_equals: expected 1 but got 0
FAIL matchMedia('(aspect-ratio: 1/1)') should not receive a change event until update the rendering step of HTML5 event loop assert_equals: expected 1 but got 0
FAIL matchMedia('(max-aspect-ratio: 4/3)') should not receive a change event until update the rendering step of HTML5 event loop assert_equals: expected 1 but got 0
FAIL matchMedia('(max-width: 150px)') should receive a change event after resize event on the window but before a requestAnimationFrame callback is called assert_array_equals: lengths differ, expected array ["resize", "change"] length 2, got [] length 0
FAIL matchMedia('(width: 100px)') should receive a change event after resize event on the window but before a requestAnimationFrame callback is called assert_array_equals: lengths differ, expected array ["resize", "change"] length 2, got [] length 0
FAIL matchMedia('(orientation: portrait)') should receive a change event after resize event on the window but before a requestAnimationFrame callback is called assert_array_equals: lengths differ, expected array ["resize", "change"] length 2, got [] length 0
FAIL matchMedia('(aspect-ratio: 1/1)') should receive a change event after resize event on the window but before a requestAnimationFrame callback is called assert_array_equals: lengths differ, expected array ["resize", "change"] length 2, got [] length 0
FAIL matchMedia('(max-aspect-ratio: 4/3)') should receive a change event after resize event on the window but before a requestAnimationFrame callback is called assert_array_equals: lengths differ, expected array ["resize", "change"] length 2, got [] length 0
Harness: the test ran to completion.
